requirements.txt
pymupdf==1.22.3
pillow==10.0.1
pytesseract==0.3.11
python-dateutil==2.8.2
pdf2image==1.16.3
